requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
babel==2.18.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
dataclasses-json==0.6.7
dateparser==1.1.8
docopt==0.6.2
editdistance==0.8.1
epitran==1.17
espeak_phonemizer==1.3.1
fastapi==0.128.0
filelock==3.20.3
Flask==3.1.2
flatbuffers==25.12.19
fsspec==2026.1.0
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fa==2.0.1
gruut_lang_fr==2.0.2
gruut_lang_it==2.0.1
gruut_lang_nl==2.0.2
gruut_lang_ru==2.0.1
gruut_lang_sw==2.0.1
h11==0.16.0
h2==4.3.0
hazm==0.7.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
Hypercorn==0.18.0
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonlines==1.2.0
libwapiti==0.2.1
llvmlite==0.46.0
marisa-trie==1.3.1
MarkupSafe==3.0.3
marshmallow==3.26.2
more-itertools==10.8.0
mpmath==1.3.0
munkres==1.1.4
mycroft_mimic3_tts==0.2.4
mypy_extensions==1.1.0
networkx==3.6.1
nltk==3.3
num2words==0.5.14
numba==0.63.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
ollama==0.6.1
onnxruntime==1.23.2
openai-whisper==20250625
packaging==26.0
pandas==3.0.0
panphon==0.22.2
phonemes2ids==1.2.2
priority==2.0.0
protobuf==6.33.5
pydantic==2.12.5
pydantic_core==2.41.5
python-crfsuite==0.9.12
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
Quart==0.20.0
quart-cors==0.8.0
regex==2026.1.15
requests==2.32.5
setuptools==80.10.2
six==1.17.0
starlette==0.50.0
swagger-ui-py==21.12.8
sympy==1.14.0
tiktoken==0.12.0
torch==2.10.0
tqdm==4.67.2
triton==3.6.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
unicodecsv==0.14.1
urllib3==2.6.3
uvicorn==0.40.0
Werkzeug==3.1.5
wsproto==1.3.2
xdgenvpy==2.4.0
